fix: ensure we receive all messages we expect
After sending a message only a single call to `conn.Receive` is made
which can cause messages to queue up in the receive buffer if not all
messages can be returned at once. This causes `ENOBUFS` as the kernel is
unable to write any further messages.

This commit introduces a check that ensures we call `conn.Receive` as
often as needed to get the right number of responses: one for
acknowledgment and one for the echo.

Resolves: #235