Stop requesting ACKs
Since https://github.com/google/nftables/pull/330 and
https://github.com/google/nftables/pull/334 were merged, we no longer
need to request an acknowledge for every message we send.

Skipping ACK requests improves runtime and memory usage for large Flush
batches, as it reduces the number of required `recvmsg` calls.

The first error will still be reported via an acknowledge even if none
are requested.

This also matches the behavior of nft userspace.